Port 11 hw/ files to C++ and add extern "C" guards to 7 headers (batch 48)
New .cpp files: hw/acpi/acpi-qmp-cmds, hw/cpu/cluster, hw/cxl/cxl-host-stubs,
hw/fsi/lbus, hw/gpio/gpio_pwr, hw/misc/arm11scu, hw/misc/armv7m_ras,
hw/nvram/chrp_nvram, hw/nvram/fw_cfg-interface, hw/tpm/tpm_ppi,
hw/xen/xen_pt_stub.

Headers with extern "C" guards: include/hw/sysbus.h,
include/hw/acpi/acpi_dev_interface.h, include/hw/cxl/cxl_host.h,
include/hw/nvram/chrp_nvram.h, include/hw/xen/xen_igd.h,
hw/tpm/tpm_ppi.h.

C++ fixes: static_cast for void* casts, TypeInfo designator reordering,
uninitialized const globals, sign comparison warnings, object_dynamic_cast
reinterpret_cast.

+++ hw/nvram/chrp_nvram.cpp
@@ -51,7 +51,7 @@ int chrp_nvram_create_system_partition(uint8_t *data, int min_len, int max_len)
     unsigned int i;
     int end;
 
-    if (max_len < sizeof(*part_header)) {
+    if (static_cast<unsigned int>(max_len) < sizeof(*part_header)) {
         goto fail;
     }
